requirements.txt
aiobotocore==2.4.2
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
boto3==1.26.98
botocore==1.29.98
certifi==2022.12.7
charset-normalizer==3.1.0
cramjam==2.6.2
fastparquet==2023.2.0
frozenlist==1.3.3
fsspec==2023.3.0
idna==3.4
jmespath==1.0.1
kafka-python==2.0.2
multidict==6.0.4
numpy==1.24.2
oauthlib==3.2.2
packaging==23.0
pandas==1.5.3
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
s3fs==2023.3.0
s3transfer==0.6.0
six==1.16.0
typing-extensions==4.5.0
urllib3==1.26.15
wrapt==1.15.0
yarl==1.8.2
